Prepare storybook folders with new page, and fix wording
