+++ _dev/apps/ui/src/router/index.ts
@@ -79,11 +79,6 @@ const routes: Array<RouteConfig> = [
     ],
   },
   {
-    path: '/product-feed/non-compliant',
-    name: 'product-feed-non-compliant',
-    component: ProductFeedPage,
-  },
-  {
     path: '/campaign',
     name: 'campaign',
     component: CampaignPage,
