+++ _dev/apps/ui/src/views/product-feed-page.vue
@@ -13,7 +13,7 @@
       v-else-if="$route.name === 'product-feed-verification-error-products'"
     />
     -->
-    <product-feed-non-compliant
+    <non-compliant-product-page
       v-else-if="$route.name === 'product-feed-verification-errors'"
     />
     <template v-else>
@@ -37,7 +37,7 @@
 import ProductFeedTableStatusDetails from '@/components/product-feed-page/product-feed-table-status-details';
 import ProductFeedPreScanTableStatusDetails from '@/components/product-feed-page/product-feed-pre-scan-table-status-details';
 import SyncOverview from '@/components/product-feed-page/dashboard/sync-overview.vue';
-import ProductFeedNonCompliant from '@/components/product-feed-page/product-feed-non-compliant/product-feed-non-compliant';
+import ProductFeedNonCompliant from '@/components/product-feed-page/non-compliant-products-page/non-compliant-products-page';
 import PsToast from '../components/commons/ps-toast';
 import {CampaignTypes} from '@/enums/reporting/CampaignStatus';
