--- _dev/apps/ui/stories/product-feed-non-compliant.stories.ts
@@ -1,17 +1,0 @@
-import ProductFeedNonCompliant from '../src/components/product-feed-page/product-feed-non-compliant.vue'
-
-export default {
-  title: 'Product feed page/Non compliant page',
-  component: ProductFeedNonCompliant,
-};
-
-const Template = (args, { argTypes }) => ({
-  props: Object.keys(argTypes),
-  components: { ProductFeedNonCompliant },
-  template: '<ProductFeedNonCompliant v-bind="$props" />',
-});
-
-export const ProductFeedNonCompliantLoading: any = Template.bind({});
-ProductFeedNonCompliantLoading.args = {
-  loading: true,
-};
